requirements.txt
backports.zoneinfo;python_version<"3.9"
django-rest-framework==0.1.0
psycopg2-binary==2.9.5
celery[redis]==5.2.7
requests==2.28.2